get rid of history uril
