+++ weave/ops_domain/run_history/run_history_v1_legacy_ops.py
@@ -120,46 +120,3 @@ def _get_history(run: wdt.Run, columns=None):
             )
             for row in history
         ]
-
-
-def _history_as_of_plugin(inputs, inner):
-    min_step = (
-        inputs.raw["asOfStep"]
-        if "asOfStep" in inputs.raw and inputs.raw["asOfStep"] != None
-        else 0
-    )
-    max_step = min_step + 1
-    alias = _make_alias(str(inputs.raw["asOfStep"]), prefix="history")
-    return f"{alias}: history(minStep: {min_step}, maxStep: {max_step})"
-
-
-def _get_history_as_of_step(run: wdt.Run, asOfStep: int):
-    alias = _make_alias(str(asOfStep), prefix="history")
-
-    data = run.gql[alias]
-    if isinstance(data, list):
-        if len(data) > 0:
-            data = data[0]
-        else:
-            data = None
-    if data is None:
-        return {}
-    return json.loads(data)
-
-
-@op(
-    render_info={"type": "function"},
-    plugins=wb_gql_op_plugin(_history_as_of_plugin),
-    hidden=True,
-)
-def _refine_history_as_of_type(run: wdt.Run, asOfStep: int) -> types.Type:
-    return wb_util.process_run_dict_type(_get_history_as_of_step(run, asOfStep))
-
-
-@op(
-    name="run-historyAsOf",
-    refine_output_type=_refine_history_as_of_type,
-    plugins=wb_gql_op_plugin(_history_as_of_plugin),
-)
-def history_as_of(run: wdt.Run, asOfStep: int) -> dict[str, typing.Any]:
-    return _get_history_as_of_step(run, asOfStep)
